feat: changed float64 to 32

+++ shared/controller.go
@@ -2,10 +2,10 @@ package shared
 
 type NormalizedGamepad struct {
 	Button_mask uint16
-	Tl          float64
-	Tr          float64
-	Lx          float64
-	Ly          float64
-	Rx          float64
-	Ry          float64
+	Tl          float32
+	Tr          float32
+	Lx          float32
+	Ly          float32
+	Rx          float32
+	Ry          float32
 }
